another small portability fix
see: https://github.com/plumed/plumed2/commit/0fee81f1062a47b767d5fc2aaddb1a4024e71e8c#commitcomment-82273698

cc: @tonigi

+++ src/tools/OpenMP.h
@@ -54,7 +54,7 @@ public:
 
 template<typename T>
 unsigned OpenMP::getGoodNumThreads(const T*x,unsigned n) {
-  unsigned long p=(unsigned long) x;
+  unsigned long long p=(unsigned long long) x;
   (void) p; // this is not to have warnings. notice that the pointer location is not used actually.
 // a factor two is necessary since there is no guarantee that x is aligned
 // to cache line boundary
